+++ internals/scripts/BabelRegister.js
@@ -1,6 +1,0 @@
-const path = require('path');
-
-require('@babel/register')({
-  extensions: ['.es6', '.es', '.jsx', '.js', '.mjs', '.ts', '.tsx'],
-  cwd: path.join(__dirname, '..', '..'),
-});