+++ src/App.tsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import icon from '../assets/icon.svg';
+
+const Hello = () => {
+  return (
+    <div>
+      <div className="Hello">
+        <img width="200px" alt="icon" src={icon} />
+      </div>
+      <h1>electron-react-boilerplate</h1>
+      <div className="Hello">
+        <a
+          href="https://electron-react-boilerplate.js.org/"
+          target="_blank"
+          rel="noreferrer"
+        >
+          <button type="button">
+            <span role="img" aria-label="books">
+              📚
+            </span>
+            Read our docs
+          </button>
+        </a>
+        <a
+          href="https://github.com/sponsors/electron-react-boilerplate"
+          target="_blank"
+          rel="noreferrer"
+        >
+          <button type="button">
+            <span role="img" aria-label="books">
+              🙏
+            </span>
+            Donate
+          </button>
+        </a>
+      </div>
+    </div>
+  );
+};
+
+export default function App() {
+  return (
+    <Router>
+      <Switch>
+        <Route path="/" component={Hello} />
+      </Switch>
+    </Router>
+  );
+}